refactor: extract attack group logic
